log ansible output
Explicitly log each ansible run to a separate file for easier
debugging. Currently, the output is *supposed* to be printing out to
STDERR, and systemd should be logging that, but something in that
process isn't working. The
[ANSIBLE_LOG_PATH](https://docs.ansible.com/ansible/2.4/config.html#envvar-ANSIBLE_LOG_PATH)
variable tells ansible to turn on logging and log to the specified
file. That should bypass whatever is going on with systemd.

Log directory set by `ANSIBLE_LOG_DIR` environment variable.

This will require that that directory exists and is writable by the
user. I will put in a separate PR to the `roles` repo to handle that.