fix subprocess pipe hanging issue
Just calling `.wait()` without reading stdout can result in the
process deadlocking when the pipe fills up.

We could call `.communicate()`, which ensures that it all gets read,
but has its own risk of memory filling up because it buffers it all:
https://docs.python.org/2/library/subprocess.html#subprocess.Popen.communicate

Instead, we just loop over it, reading line by line until it hits a
`StopIteration`.

I do not have a unit test for this bug, unfortunately. It's a tricky
interaction between python and the system that it's running on and
would likely be a flakey test at best. Furthermore, reproducing the
bug and successfully testing for it would involve detecting that the
program had hung, and that gets tricky (halting problem and all).